Move SendableResponse from leshan-server-core to leshan-core
